pyproject.toml
[tool.poetry]
name = "infertalent-api"
version = "0.1.0"
description = "infertalents api"
authors = ["none"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-simple-history = "^3.0.0"
django-cors-headers = "^3.10.0"
psycopg2-binary = "^2.9.2"
django-extensions = "^3.1.5"
pyotp = "^2.6.0"
django-ses = "^2.3.0"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "4.4.0"
html2text = "^2020.1.16"
phonenumbers = "^8.12.37"
django-autocomplete-light = "^3.8.2"
boto3 = "^1.20.23"
django-storages = "^1.12.3"
pandas = "^1.4.1"
cryptography = "^36.0.0"
channels = "^3.0.4"
affinda = "^1.0.1"
PyJWT = "1.7.1"
geopy = "^2.3.0"
mozilla-django-oidc = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
pre-commit = "^2.15.0"
pre-commit-hooks = "^4.0.1"
isort = "^5.10.1"
flake8 = "^4.0.1"
tox = "^3.24.4"
pygraphviz = "^1.7"
coverage = "^6.3.2"

[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
